Make version 1.0.0 to indicate stable release
This project has already been used in production for some time,
so backwards-compatibility is already important. The project
also claims to follow semver, but semver doesn't have any impact
until version 1.0.0.  Per [1], "If your software is being used in
production, it should probably already be 1.0.0."

"Used by production service" and "this project is alpha,
version 0.x and we can change anything at any time" are not
compatible; it is time to take stability seriously. Mark the
version as 1.0.0 accordingly.

[1] https://semver.org/spec/v2.0.0.html