docs: fix weird formatting in changelog
In this format of changelog, the purpose of the [brackets] around
version numbers is that they'll be rendered as hyperlinks; see
for example another project using this format at [1].

However, this only works if you define the targets of the hyperlinks
(e.g. at end of the file). Since this was not done for iiblib's
changelog, the versions don't render as hyperlinks and the brackets
just make the log look more noisy - so let's not do that.

[1] https://github.com/release-engineering/pushsource/blob/master/CHANGELOG.md