Fix rendering on PyPI by using markdown for README
If you have a look at entries for this project on PyPI, such
as [1], you'll see that there's an ugly mix of formatting.
The README sections are formatted nicely, but the changelog
shows as unrendered markdown.

The reason for this is that we're concatening README.rst and
CHANGELOG.md into one string for the long description. Although
pypi does support rendering both formats, it's not going to
figure out that the format changes from RST to markdown halfway
through the document...

We can make it render nicely, but we have to pick one format.
Let's do that by moving the README to markdown.

[1] https://pypi.org/project/iiblib/1.0.0/