Remove retries based on specific state_reason
The retries added in [1] don't work because the retried methods
don't wait for IIB build to finish, which means that they can never
encounter an error state. The blocking method, "wait_for_build" has a
side effect of raising IIBRecoverableError exception if a retriable
error is found. This is undesirable because no exception should be
raised if an IIB build fails.

This commit removes the nonfunctional retries and makes sure that no
unnecessary exception is raised. The retry mechanism will not be
re-implemented anywhere else because the negotiation token cannot be
reused in multiple IIB requests.

[1] https://github.com/release-engineering/iiblib/pull/57